requirements.txt
Flask==3.0.0
Telethon==1.34.0
python-dotenv==1.0.0
gunicorn==21.2.0
